deleted old services and models, cleaned up configuration files, and added/edited docs and readme
