+++ app/config/database.py
@@ -1,0 +1,10 @@
+class DB:
+    CONFIG = {
+        "dialect" : "mysql",
+        "driver" : "pymysql",
+        "host" : "db",
+        "user" : "user",
+        "password" : "pass",
+        "database" : "payroll_db",
+        "port" : 3306
+    }
