+++ app/extensions.py
@@ -1,6 +1,5 @@
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
-
 db = SQLAlchemy()
 migrate = Migrate()
