--- bak/config/__init__.py
@@ -1,9 +1,0 @@
-from .config import (
-    DATABASE_CONFIG,
-    CONNECTION_ATTEMPTS,
-    RECONNECTION_DELAY,
-    LOG_FORMATTER,
-    LOG_PATH,
-    LOG_ROTATION,
-    LOG_LEVEL,
-)
