--- bak/routes/__init__.py
@@ -1,6 +1,0 @@
-from flask import Blueprint
-
-from .roles_routes import roles_bp
-
-def register_routes(app):
-    app.register_blueprint(roles_bp)
